+++ src/main/java/com/pavan/rest/controller/FdController.java
@@ -10,12 +10,15 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.pavan.beans.ApiResponse;
 import com.pavan.beans.FdBean;
+import com.pavan.modal.User;
 import com.pavan.service.FdService;
+import com.pavan.service.LoginService;
 import com.pavan.util.Utility;
 
 @RestController
@@ -24,23 +27,33 @@ import com.pavan.util.Utility;
 public class FdController {
 
 	@Autowired
+	private LoginService loginService;
+
+	@Autowired
 	private FdService fdService;
+
 	private String message;
 
 	@PostMapping
 	@Transactional(rollbackOn = Exception.class)
-	public ApiResponse saveFd(@RequestBody(required = true) FdBean fdBean) {
+	public ApiResponse saveFd(@RequestHeader("userToken") String userToken,
+			@RequestBody(required = true) FdBean fdBean) {
 
 		try {
-			fdService.saveFd(fdBean);
+			User currentUser = loginService.getUserFromToken(userToken);
+			if (currentUser != null) {
+				fdService.saveFd(fdBean, currentUser);
 
-			if (Utility.isEmpty(fdBean.getId())) {
-				message = "Fd saved successfully";
+				if (Utility.isEmpty(fdBean.getId())) {
+					message = "Fd saved successfully";
+				} else {
+					message = "Fd updated successfully";
+				}
+
+				return new ApiResponse(HttpStatus.OK, message, null);
 			} else {
-				message = "Fd updated successfully";
+				return new ApiResponse(HttpStatus.UNAUTHORIZED, "Unauthorized to acccess this resource", null);
 			}
-
-			return new ApiResponse(HttpStatus.OK, message, null);
 		} catch (Exception e) {
 			message = "Error-" + e.getMessage();
 			return new ApiResponse(HttpStatus.INTERNAL_SERVER_ERROR, message, null);
@@ -49,22 +62,52 @@ public class FdController {
 	}
 
 	@GetMapping
-	public ApiResponse fds() {
-		return fdService.getFds();
+	public ApiResponse fds(@RequestHeader("userToken") String userToken) {
+		User currentUser = loginService.getUserFromToken(userToken);
+		if (currentUser != null) {
+			return fdService.getFds(currentUser);
+		} else {
+			return new ApiResponse(HttpStatus.UNAUTHORIZED, "Unauthorized to acccess this resource", null);
+		}
 	}
 
 	@GetMapping("/{id}")
-	public ApiResponse getFd(@PathVariable(value = "id") Long id) {
-		return fdService.getFd(id);
+	public ApiResponse getFd(@RequestHeader("userToken") String userToken, @PathVariable(value = "id") Long id) {
+		User currentUser = loginService.getUserFromToken(userToken);
+		if (currentUser != null) {
+			return fdService.getFd(id);
+		} else {
+			return new ApiResponse(HttpStatus.UNAUTHORIZED, "Unauthorized to acccess this resource", null);
+		}
 	}
 
 	@DeleteMapping("/{id}")
-	public ApiResponse deleteFd(@PathVariable(value = "id") Long id) {
-		return fdService.deleteFd(id);
+	public ApiResponse deleteFd(@RequestHeader("userToken") String userToken, @PathVariable(value = "id") Long id) {
+		User currentUser = loginService.getUserFromToken(userToken);
+		if (currentUser != null) {
+			return fdService.deleteFd(id);
+		} else {
+			return new ApiResponse(HttpStatus.UNAUTHORIZED, "Unauthorized to acccess this resource", null);
+		}
 	}
 
-	@DeleteMapping()
-	public ApiResponse deleteFds() {
-		return fdService.deleteFds();
+	@DeleteMapping
+	public ApiResponse deleteFds(@RequestHeader("userToken") String userToken) {
+		User currentUser = loginService.getUserFromToken(userToken);
+		if (currentUser != null) {
+			return fdService.deleteFds(currentUser);
+		} else {
+			return new ApiResponse(HttpStatus.UNAUTHORIZED, "Unauthorized to acccess this resource", null);
+		}
+	}
+	
+	@DeleteMapping("/all")
+	public ApiResponse deleteAllFds(@RequestHeader("userToken") String userToken) {
+		User currentUser = loginService.getUserFromToken(userToken);
+		if (currentUser != null) {
+			return fdService.deleteFds(null);
+		} else {
+			return new ApiResponse(HttpStatus.UNAUTHORIZED, "Unauthorized to acccess this resource", null);
+		}
 	}
 }
